srv: bugfix write -> create new file

+++ server/modules/rm3json.py
@@ -82,7 +82,8 @@ def write(file, data, call_from=""):
     with open(file3, 'wb') as outfile:
         json.dump(data, codecs.getwriter('utf-8')(outfile), ensure_ascii=False, sort_keys=True, indent=4)
 
-    os.remove(file2)
+    if os.path.isfile(file2):
+        os.remove(file2)
     os.rename(file3, file2)
 
     json_logging.debug("write ... " + file + " ...from:" + call_from)
